requirements.txt
requests
pandas
scikit-learn
xgboost
streamlit
python-dotenv
joblib
